rename hetnode2vec, remove gamma and all method related to hetrogenous graph
